+++ Classes/DataHandling/DataHandler.php
@@ -2,8 +2,7 @@
 
 namespace SourceBroker\Hugo\DataHandling;
 
-use SourceBroker\Hugo\Service\ExportContentService;
-use SourceBroker\Hugo\Service\ExportPageService;
+use SourceBroker\Hugo\Queue\QueueInterface;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\MathUtility;
@@ -16,18 +15,19 @@ class DataHandler implements SingletonInterface
      * Clears path and URL caches if the page was deleted.
      *
      * @param string $tableName
-     * @param string|int $id
+     * @param string|int $recordId
      * @throws \TYPO3\CMS\Core\Locking\Exception\LockAcquireException
      * @throws \TYPO3\CMS\Core\Locking\Exception\LockCreateException
      */
-    public function processCmdmap_deleteAction($tableName, $id)
+    public function processCmdmap_deleteAction($tableName, $recordId)
     {
-        // TODO: optimize later
-        if ($tableName === 'tt_content') {
-            $this->deleteHugoContentElements((int)$id);
+	$queueValue = 'delete:'.$tableName;
+
+        if (is_numeric($recordId)) {
+            $queueValue .= ':'.intval($recordId);
         }
-        $this->exportHugoPages();
-    }
+
+        $this->getQueue()->push($queueValue);    }
 
     /**
      * Expires caches if the page was moved.
@@ -43,19 +43,13 @@ class DataHandler implements SingletonInterface
         $tableName,
         $recordId
     ) {
-        if ($command === 'undelete') {
-            switch ($tableName) {
-                case 'pages':
-                    $this->exportHugoPages();
-                    break;
-                case 'tt_content':
-                    $this->exportHugoContentElements($recordId);
-                    $this->exportHugoPages();
-                    break;
-            }
-        } else {
-            $this->exportHugoPages();
+        $queueValue = 'update:'.$tableName;
+
+        if (is_numeric($recordId)) {
+            $queueValue .= ':'.$recordId;
         }
+
+        $this->getQueue()->push($queueValue);
     }
 
     /**
@@ -79,52 +73,25 @@ class DataHandler implements SingletonInterface
         /** @noinspection PhpUnusedParameterInspection */
         \TYPO3\CMS\Core\DataHandling\DataHandler $dataHandler
     ) {
+        $queueValue = 'update:'.$tableName;
+
         if (!MathUtility::canBeInterpretedAsInteger($recordId)) {
             $recordId = (int)$dataHandler->substNEWwithIDs[$recordId];
         }
-        switch ($tableName) {
-            case 'pages':
-                $this->exportHugoPages();
-                break;
-            case 'tt_content':
-                $this->exportHugoContentElements($recordId);
-                $this->exportHugoPages();
-                break;
+
+        if (is_numeric($recordId)) {
+            $queueValue .= ':'.$recordId;
         }
+
+        $this->getQueue()->push($queueValue);
     }
 
     /**
-     * @throws \TYPO3\CMS\Core\Locking\Exception\LockAcquireException
-     * @throws \TYPO3\CMS\Core\Locking\Exception\LockCreateException
+     * @return QueueInterface
      */
-    public function exportHugoPages()
+    public function getQueue(): QueueInterface
     {
         $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $hugoExportPageService = $objectManager->get(ExportPageService::class);
-        $hugoExportPageService->exportAll();
-    }
-
-    /**
-     * @param null $contentRecordUid
-     *
-     * @throws \TYPO3\CMS\Core\Locking\Exception\LockAcquireException
-     * @throws \TYPO3\CMS\Core\Locking\Exception\LockCreateException
-     */
-    public function exportHugoContentElements($contentRecordUid = null)
-    {
-        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $hugoExportContentService = $objectManager->get(ExportContentService::class);
-        if ($contentRecordUid === null) {
-            $hugoExportContentService->exportAll();
-        } else {
-            $hugoExportContentService->exportSingle($contentRecordUid);
-        }
-    }
-
-    protected function deleteHugoContentElements(int $contentRecordUid): void
-    {
-        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $hugoExportContentService = $objectManager->get(ExportContentService::class);
-        $hugoExportContentService->deleteSingle($contentRecordUid);
+        return $objectManager->get(QueueInterface::class);
     }
 }