+++ ext_tables.sql
@@ -5,3 +5,16 @@ CREATE TABLE tx_dce_domain_model_dce (
 CREATE TABLE sys_domain (
 	tx_hugo_domains varchar(255),
 );
+
+
+CREATE TABLE tx_hugo_queueitems (
+  uid int(11) NOT NULL auto_increment,
+  namespace varchar (32) NOT NULL,
+  value varchar(128) NOT NULL,
+  executed tinyint(4) NOT NULL DEFAULT '0',
+  created_date DATETIME,
+  executed_date DATETIME,
+  PRIMARY KEY (uid),
+  KEY namespace (namespace, value),
+  KEY executed (executed, created_date),
+);